// backend/src/common/enums/collection-activity-type.enum.ts
export enum CollectionActivityType {
  PHONE_CALL = 'Phone Call',
  EMAIL_SENT = 'Email Sent',
  SMS_SENT = 'SMS Sent',
  LETTER_SENT = 'Letter Sent',
  IN_PERSON_VISIT = 'In Person Visit',
  PROMISE_TO_PAY = 'Promise to Pay',
  PAYMENT_ARRANGEMENT = 'Payment Arrangement',
  RIGHT_PARTY_CONTACT = 'Right Party Contact',
  CEASE_AND_DESIST = 'Cease and Desist',
  SKIP_TRACE = 'Skip Trace',
}